+++ src/XYIndex.h
@@ -6,7 +6,7 @@
 #define NUM_ROWS 39
 
 // function to map a x, y coordinate to the index into the led array
-uint16_t XYToIndex(uint8_t x, uint8_t y);
+uint16_t XYToIndex(uint16_t x, uint16_t y);
 
 #ifdef DEBUG
 // loop through all pixels via x,y coordinates making sure they all get mapped correctly
